FEAT: socket refined
