+++ front/src/components/WebGL/function/update.ts
@@ -2,14 +2,15 @@ import data from '../interface/gameData';
 import PhysicsEngine from "../class/PhysicsEngine";
 import { GameManager } from "../class/GameManager";
 import { AIManager } from "../class/AIManager";
+import { ItemManager } from '../class/ItemManager';
 
 function update(delta: number) {
-	// if (data.mode === 'object') {
+	if (data.mode === 'object') {
 	// 	/* 아이템 생성 */
-	// 	ItemManager.getInstance().createItem();
+		ItemManager.getInstance().createItem();
 	// 	/* 아이템 업데이트 */
-	// 	ItemManager.getInstance().updateItems(delta);
-	// }
+		ItemManager.getInstance().updateItems(delta);
+	}
 	/* 공 위치 업데이트 */
 	PhysicsEngine.GuaranteeConflict(data.ball, delta);
 	/* 게임 종료 조건 확인 */
